Move base User specs into their own module
